CNVcalling Runs a CNV analysis given a seurat object

CNVcalling Runs a CNV analysis given a seurat object

CNVcalling(
  seuratObj,
  assay = NULL,
  referenceVar = NULL,
  referenceLabel = NULL,
  scaleOnReferenceLabel = TRUE,
  thresholdPercentile = 0.01,
  genes = getGenes(),
  windowSize = 100,
  windowStep = 20,
  topNGenes = 7000
)

Arguments

seuratObj

The Seurat object containing the data to analyze

assay

Name of the assay to run the CNV on. Takes the results of prepareCountsForCNVAnalysis by default if available

referenceVar

The variable name of the annotations in the Seurat metadata

referenceLabel

The label given to the observations wanted as reference (can be any type of annotation)

scaleOnReferenceLabel

If you want to scale the results depending on the reference observations

thresholdPercentile

which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.

genes

List of genes from ensembl

windowSize

Size of the genomic windows

windowStep

Step between the genomic windows

topNGenes

Number of top expressed genes to keep

Value

This function returns the genomic scores per genomic window