A function to plot the CNV results into a heatmap

A function to plot the CNV results into a heatmap

Usage

plotCNVResults(
  seuratObj,
  referenceVar = NULL,
  splitPlotOnVar = referenceVar,
  savePath = ".",
  downsizePlot = FALSE
)

Arguments

seuratObj

Seurat object containing the data used to get the genomicScores.

referenceVar

The variable name of the annotations in the Seurat metadata

splitPlotOnVar

The variable name on which to split the heatmap rows.

savePath

Path to save the pdf heatmap. If NULL, plot won't be saved (default = .).

downsizePlot

Subset the observations to speed up the plotting process (default = FALSE).

Value

This function builds a heatmap and saves it as a .pdf file in your working directory