CNVanalysis Runs the CNV functions given a seurat object or a list of seurat objects

CNVanalysis Runs the CNV functions given a seurat object or a list of seurat objects

Usage

CNVanalysis(
  object,
  referenceVar = NULL,
  referenceLabel = NULL,
  doRecapPlot = TRUE,
  pooledReference = TRUE,
  scaleOnReferenceLabel = TRUE,
  assay = NULL,
  thresholdPercentile = 0.01,
  genes = getGenes(),
  windowSize = 100,
  windowStep = 20,
  topNGenes = 7000
)

Arguments

object

The Seurat object or list of seurat objects containing the data to analyze

referenceVar

The variable name of the annotations in the Seurat metadata

referenceLabel

The label given to the observations wanted as reference (can be any type of annotation)

doRecapPlot

Default TRUE. Will output the CNV heatmaps by annotation if TRUE.

pooledReference

Default TRUE. Will build a pooled reference across all samples if TRUE.

scaleOnReferenceLabel

If you want to scale the results depending on the normal observations

assay

Name of the assay to run the CNV on. Takes the results of prepareCountsForCNVAnalysis by default if available

thresholdPercentile

Which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.

genes

List of genes from ensembl

windowSize

Size of the genomic windows

windowStep

Step between the genomic windows

topNGenes

Number of top expressed genes to keep

Value

This function returns the genomic scores per genomic window per seurat object. If given a list with more than one seurat object, and their annotations, it will output a heatmap per cell type given