CNVcalling Runs a CNV analysis given a seurat object

CNVcalling Runs a CNV analysis given a seurat object

Usage

CNVcalling(
  seuratObj,
  assay = NULL,
  referenceVar = NULL,
  referenceLabel = NULL,
  scaleOnReferenceLabel = TRUE,
  denoise = TRUE,
  thresholdPercentile = 0.01,
  geneMetadata = getGenes(),
  genesToForce = NULL,
  windowSize = 150,
  windowStep = 10,
  saveGenomicWindows = FALSE,
  topNGenes = 7000
)

Arguments

seuratObj

The Seurat object containing the data to analyze

assay

Name of the assay to run the CNV on. Takes the results of prepareCountsForCNVAnalysis by default if available

referenceVar

The variable name of the annotations in the Seurat metadata

referenceLabel

The label given to the observations wanted as reference (can be any type of annotation)

scaleOnReferenceLabel

If you want to scale the results depending on the reference observations

denoise

If the data needs to be denoised (default = TRUE)

thresholdPercentile

which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.

geneMetadata

Dataframe of genes and their metadata from ensembl

genesToForce

List of genes to force into the analysis (by default, fastCNV only takes the topNgenes most expressed).

windowSize

Size of the genomic windows

windowStep

Step between the genomic windows

saveGenomicWindows

If the information of the genomic windows need to be saved in the current directory (default = FALSE).

topNGenes

Number of top expressed genes to keep

Value

This function returns the genomic scores per genomic window