fastCNV calls all of the internal functions needed to compute the putative CNV on a seurat object or a list of seurat objects

fastCNV calls all of the internal functions needed to compute the putative CNV on a seurat object or a list of seurat objects

Usage

fastCNV(
  seuratObj,
  sampleName,
  referenceVar = NULL,
  referenceLabel = NULL,
  splitPlotOnVar = referenceVar,
  downsizePlot = FALSE,
  doPlot = TRUE,
  doRecapPlot = TRUE,
  prepareCounts = TRUE,
  pooledReference = TRUE,
  assay = NULL,
  getCNVPerChromosomeArm = TRUE,
  savePath = ".",
  aggregFactor = 15000,
  seuratClusterResolution = 0.8,
  aggregateByVar = T,
  reClusterSeurat = F,
  scaleOnReferenceLabel = TRUE,
  thresholdPercentile = 0.01,
  genes = getGenes(),
  windowSize = 100,
  windowStep = 20,
  topNGenes = 7000
)

Arguments

seuratObj

Seurat object or list of seurat objects

sampleName

Name of the sample or list of names corresponding to the samples in the seuratObj

referenceVar

The variable name of the annotations in the Seurat metadata

referenceLabel

The label given to the observations wanted as reference (can be any type of annotation)

splitPlotOnVar

The name given to the meta.data slot to split the observations on during the plotCNVResults step, if different from referenceVar.

downsizePlot

Subset the observations to speed up the plotting process (default = FALSE).

doPlot

If TRUE will plot a heatmap for each of the samples (default = TRUE)

doRecapPlot

Default TRUE. Will output the CNV heatmaps by annotation if TRUE.

prepareCounts

If FALSE will not run the prepareCountsForCNVAnalysis function. Default TRUE.

pooledReference

Default TRUE. Will build a pooled reference across all samples if TRUE.

assay

Name of the assay to run the CNV on. Takes the results of prepareCountsForCNVAnalysis by default if available

getCNVPerChromosomeArm

If TRUE, will save into the metadata the CNV per chromosome arm

savePath

Default .. If NULL the heatmap won't be saved as a pdf.

aggregFactor

The number of counts per spot desired (default 30 000). If less than 1000, will not run the prepareCountsForCNVAnalysis function.

seuratClusterResolution

The resolution wanted for the seurat clusters (default 0.8)

aggregateByVar

If referenceVar is given, whether to use it to pool the observations

reClusterSeurat

Whether to re-cluster if the Seurat object given already has a seurat_clusters slot in its meta.data

scaleOnReferenceLabel

If you want to scale the results depending on the normal observations

thresholdPercentile

Which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.

genes

List of genes (default genes from ensembl version 111)

windowSize

Size of the genomic windows

windowStep

Step between the genomic windows

topNGenes

Number of top expressed genes to keep

Value

This function returns a list of the seurat objects after all the analysis and it creates the heatmaps of the CNVs of every object in seuratObj and saves them in a pdf in the current working directory