A function to aggregate observations with the same cell types to get higher counts per observation. It will help to have higher counts to compute the CNV

A function to aggregate observations with the same cell types to get higher counts per observation. It will help to have higher counts to compute the CNV

Usage

prepareCountsForCNVAnalysis(
  seuratObj,
  sampleName = NULL,
  referenceVar = NULL,
  aggregateByVar = T,
  aggregFactor = 15000,
  seuratClusterResolution = 0.8,
  reClusterSeurat = F
)

Arguments

seuratObj

Seurat object containing the data

sampleName

Name of the sample

referenceVar

The variable name of the annotations in the Seurat metadata

aggregateByVar

If a referenceVar is given, whether to use it to aggregate the observations depending on their type. Default TRUE.

aggregFactor

The number of counts per observation desired (default 30 000)

seuratClusterResolution

The resolution wanted for the seurat clusters (default 0.8)

reClusterSeurat

If TRUE, clustering is re-done for the Seurat object.

Value

This function returns a Seurat object with the modified count matrix in a new assay called AgregatedCounts and the Seurat clusters in the metadata