Aggregate Observations by Cell Type for CNV Analysis Aggregates observations with the same cell types to increase counts per observation, improving Copy Number Variation (CNV) computation.

Aggregate Observations by Cell Type for CNV Analysis Aggregates observations with the same cell types to increase counts per observation, improving Copy Number Variation (CNV) computation.

Usage

prepareCountsForCNVAnalysis(
  seuratObj,
  sampleName = NULL,
  referenceVar = NULL,
  aggregateByVar = T,
  aggregFactor = 15000,
  clusterResolution = 0.8,
  reCluster = F
)

Arguments

seuratObj

A Seurat object containing the data.

sampleName

A character string specifying the sample name.

referenceVar

The name of the metadata column in the Seurat object that contains reference annotations.

aggregateByVar

Logical. If TRUE (default), aggregates observations based on referenceVar annotations.

aggregFactor

Integer. The target number of counts per observation (default = 15 000).

clusterResolution

Numeric. The resolution used for Seurat clustering (default = 0.8).

reCluster

Logical. If TRUE, re-runs clustering on the Seurat object.

Value

A Seurat object with:

A new assay called "AggregatedCounts" containing the modified count matrix.

Seurat clusters stored in the metadata.